+++ app/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask_cors import CORS
 
-from app.config.logging_config import setup_logging
+from app.config.logging import setup_logging
 from app.api.routes import api, init_routes
 
 def create_app(wifi_monitor=None):
@@ -19,5 +19,5 @@ def create_app(wifi_monitor=None):
     # Initialize routes with monitor if provided
     if wifi_monitor:
         init_routes(wifi_monitor)
-    
+
     return app
